move `DynLib` & `DxgiLib` into auxil
