+++ wgpu-hal/src/auxil/dxgi/mod.rs
@@ -3,6 +3,7 @@
 // Vulkan-on-Windows backend also uses.
 #[cfg(dx12)]
 pub mod conv;
+pub mod dxgi_lib;
 #[cfg(dx12)]
 pub mod exception;
 #[cfg(dx12)]
@@ -10,7 +11,6 @@ pub mod factory;
 pub mod hdr;
 #[cfg(dx12)]
 pub mod name;
-#[cfg(dx12)]
 pub mod result;
 #[cfg(dx12)]
 pub mod time;
